cloud_functions/handle_messages/requirements.txt
slack-sdk==3.21.0
google-cloud-pubsub==2.10.0
requests>=2.0
google-cloud-firestore==2.11.0
redis